chain: move BlockHeight into block
