+++ zebra-chain/src/parameters/tests.rs
@@ -3,7 +3,7 @@
 use super::*;
 use NetworkUpgrade::*;
 
-use crate::types::BlockHeight;
+use crate::block::BlockHeight;
 use crate::{Network, Network::*};
 
 use std::collections::HashSet;
