+++ zebra-consensus/src/chain.rs
@@ -27,8 +27,8 @@ use std::{
 use tower::{buffer::Buffer, Service, ServiceExt};
 use tracing_futures::Instrument;
 
+use zebra_chain::block::BlockHeight;
 use zebra_chain::block::{Block, BlockHeaderHash};
-use zebra_chain::types::BlockHeight;
 use zebra_chain::{Network, NetworkUpgrade::Sapling};
 
 /// The maximum expected gap between blocks.
